Refactor: +BoardPanel:move all the game logic to BoardPanel class +MineSweeper:MineSweeper class now only works with Panel +SettingPanel: Add logic to set difficulties Add: +SettingMenu class
