+++ BoardPanel.java
@@ -1,22 +1,33 @@
 import java.awt.*;
 import java.awt.event.*;
 import java.util.ArrayList;
+import java.util.Random;
 import javax.swing.*;
 
 public class BoardPanel extends JPanel {
-    private MineSweeper mineSweeper;
+    private MineSweeper game;
     private int numRows;
     private int numCols;
+    // 2d array to store each mine tile so that we know where each is
     private MineTile[][] board;
     private ArrayList<MineTile> mineList;
 
-    public BoardPanel(MineSweeper mineSweeper, int numRows, int numCols, MineTile[][] board, ArrayList<MineTile> mineList) {
-        this.mineSweeper = mineSweeper;
+    private int mineCount ;
+
+    Random random = new Random();
+
+    int tilesClicked = 0;
+    boolean gameOver = false;
+
+    public BoardPanel(MineSweeper mineSweeper, int numRows, int numCols,int mineCount) {
+        this.game = mineSweeper;
         this.numRows = numRows;
         this.numCols = numCols;
-        this.board = board;
-        this.mineList = mineList;
+        this.board = new MineTile[numRows][numCols];
+        this.mineList = new ArrayList<>();
+        this.mineCount=mineCount;
 
+        game.setTextPanel("MineSweeper");
         setLayout(new GridLayout(numRows, numCols));
         setBackground(Color.pink);
 
@@ -32,22 +43,22 @@ public class BoardPanel extends JPanel {
                 tile.addMouseListener(new MouseAdapter() {
                     @Override
                     public void mousePressed(MouseEvent e) {
-                        if (mineSweeper.gameOver) {
+                        if (gameOver) {
                             return;
                         }
                         MineTile tile = (MineTile) e.getSource();
 
                         if (e.getButton() == MouseEvent.BUTTON1) {
                             if (tile.getText().equals("")) {
-                                if (mineSweeper.tilesClicked == 0) {
+                                if (tilesClicked == 0) {
                                     if (mineList.contains(tile)) {
-                                        mineSweeper.setMines();
+                                        setMines();
                                     }
                                 }
                                 if (mineList.contains(tile)) {
-                                    mineSweeper.revealMines();
+                                    revealMines();
                                 } else {
-                                    mineSweeper.checkMine(tile.r, tile.c);
+                                    checkMine(tile.r, tile.c);
                                 }
                             }
                         } else if (e.getButton() == MouseEvent.BUTTON3) {
@@ -62,5 +73,83 @@ public class BoardPanel extends JPanel {
                 add(tile);
             }
         }
+        setMines();
+    }
+
+    void setMines() {
+        mineList.clear();
+        int mineLeft = mineCount;
+        while (mineLeft > 0) {
+            int r = random.nextInt(numRows);
+            int c = random.nextInt(numCols);
+
+            MineTile tile = board[r][c];
+            if (!mineList.contains(tile)) {
+                mineList.add(tile);
+                mineLeft -= 1;
+            }
+        }
+    }
+
+    void revealMines() {
+        for (MineTile tile : mineList) {
+            tile.setText("💣");
+        }
+        gameOver = true;
+        game.setTextPanel("Game Over");
+    }
+
+    void checkMine(int r, int c) {
+        if (r < 0 || r >= numRows || c < 0 || c >= numCols) {
+            return;
+        }
+
+        MineTile tile = board[r][c];
+        if (!tile.isEnabled()) {
+            return;
+        }
+
+        tile.setEnabled(false);
+        tilesClicked += 1;
+
+        int minesFound = 0;
+        minesFound += countMine(r - 1, c - 1);
+        minesFound += countMine(r - 1, c);
+        minesFound += countMine(r - 1, c + 1);
+        minesFound += countMine(r, c - 1);
+        minesFound += countMine(r, c + 1);
+        minesFound += countMine(r + 1, c - 1);
+        minesFound += countMine(r + 1, c);
+        minesFound += countMine(r + 1, c + 1);
+
+        if (minesFound > 0) {
+            tile.setText(Integer.toString(minesFound));
+        } else {
+            tile.setText("");
+            checkMine(r - 1, c - 1);
+            checkMine(r - 1, c);
+            checkMine(r - 1, c + 1);
+            checkMine(r, c - 1);
+            checkMine(r, c + 1);
+            checkMine(r + 1, c - 1);
+            checkMine(r + 1, c);
+            checkMine(r + 1, c + 1);
+        }
+
+        if (tilesClicked == numRows * numCols - mineList.size()) {
+            gameOver = true;
+            game.setTextPanel(
+                    "Mines Cleared!");
+        }
+    }
+
+    int countMine(int r, int c) {
+        if (r < 0 || r >= numRows || c < 0 || c >= numCols) {
+            return 0;
+        }
+        if (mineList.contains(board[r][c])) {
+            return 1;
+        }
+        return 0;
     }
 }
